Add MinGW support

+++ tests/ringbuf.cpp
@@ -14,19 +14,19 @@ protected:
         delete[] buffer;
     }
 
-    void initRingBuf(uint size)
+    void initRingBuf(size_t size)
     {
         buffer = new uint8_t[size + 10];
         memset(buffer, 0, size + 10);
         ringbuf_init(&_ringbuf, buffer + 5, size);
     }
 
-    void append(void* data, uint size)
+    void append(void* data, size_t size)
     {
         ringbuf_write(&_ringbuf, data, size);
     }
 
-    void erase(uint size)
+    void erase(size_t size)
     {
         ringbuf_erase(&_ringbuf, size);
     }
@@ -41,7 +41,7 @@ protected:
         return ringbuf_read_uint8(&_ringbuf);
     }
 
-    void expectFreeSpace(uint freeSpace)
+    void expectFreeSpace(size_t freeSpace)
     {
         EXPECT_EQ(ringbuf_get_free_space(&_ringbuf), freeSpace);
     }
@@ -56,7 +56,7 @@ protected:
         EXPECT_TRUE(ringbuf_is_empty(&_ringbuf));
     }
 
-    void expectData(void* expected, uint size)
+    void expectData(void* expected, size_t size)
     {
         uint8_t temp[size];
         ringbuf_peek(&_ringbuf, temp, size, 0);
